requirements.txt
ultralytics>=8.0
opencv-python>=4.0
pillow>=10.0
numpy>=1.24
matplotlib>=3.7
torch>=2.0